Major change trees must be of type tree :: { any, [tree] }
This changes the supported tree data type from the old [ value |
children ] to { any, [ tree ] }. As suggested in the previous commit
this change allows guards to be written using tuple accessors. Moreover
the old data type was ambigious at best as to whether a location in the
tree was a node or leaf. This now explicity supports nodes that have a
value in addition to child nodes.