change supported data type back to what it was
... not using `tree :: { any, [tree] }` any more, returning back to old
`tree :: [ any | [tree] ] | leaf` defintion so that it's more like
existing implementations like that of clojure.
Major difference from the original version now is that in any node, the
*first* value is assumed to be the value of this node. What this means
is that calling `down` on a node will no long descend into the first
value and it will be added to the context as the value of the current
parent.
Eg.
current: [ 1, 2, 3, [4] ] -down-> current: 2, right: [3, [4]]